// synara/src/app/features/matrix-dto/generated/DirectoryRoomHitDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DirectoryRoomHitDto = { roomId: string, name?: string, topic?: string, canonicalAlias?: string, avatarUrl?: string, memberCount: number, worldReadable: boolean, guestCanJoin: boolean, roomType: "room" | "space", };