// synara/src/app/features/matrix-dto/generated/NativeRoomDirectoryPage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DirectoryRoomHitDto } from "./DirectoryRoomHitDto";

export type NativeRoomDirectoryPage = { sessionGeneration: number, requestId: number, chunk: Array<DirectoryRoomHitDto>, prevBatch?: string, nextBatch?: string, };