// synara/src/app/features/matrix-dto/generated/NativeRoomDirectoryProtocols.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DirectoryProtocolInstance } from "./DirectoryProtocolInstance";

export type NativeRoomDirectoryProtocols = { sessionGeneration: number, instances: Array<DirectoryProtocolInstance>, };